pyproject.toml
[project]
name = "gemini-vision-mcp"
version = "0.1.0"
description = "Visual regression MCP server with Gemini 3 Flash agentic vision"
authors = [
    {name = "Bret Bouchard", email = "yesreply@happy.engineering"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.26.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "opencv-python>=4.10.0",
    "mss>=9.0.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wheres_waldo"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
